build(devcontainer): Enable markdown linting, add hatch, and update extensions EL-1009
Enable the installation of markdownlint-cli in the dev container to ensure compliance with markdown formatting guidelines.

Add the Hatch build tool to the development dependencies to adhere to the project's development policy for packaging and environment management.

Update the recommended VS Code extensions for the development container, replacing the print extension with the Edge DevTools extension.

Sunday 2025-04-27 15:51:33 Vancouver GMT-07:00